Refactor extract client tests to package main(client_test)
They are more standalone client tests than server tests.
